feat!: Migrate to support ESLint v9
This is breaking as internal API was changed by ESLint for plugins.
* This now also provides flat config configurations
* ESLint no longer supports environments, so this was replaced by
  `@nextcloud/recommended-globals` config
