--- .travis.yml
@@ -1,5 +1,0 @@
-language: node_js
-node_js: node
-branches:
-  only:
-  - master
