+++ lib/configs/index.ts
@@ -1,0 +1,4 @@
+import { recommended } from './recommended'
+import { recommendedWithGlobals } from './recommended-globals'
+
+export default [recommended, recommendedWithGlobals]
