--- lib/environments/nextcloud.js
@@ -1,11 +1,0 @@
-"use strict";
-
-module.exports = {
-	globals: {
-		t: 'readonly',
-		n: 'readonly',
-		OC: 'readonly',
-		OCA: 'readonly',
-		OCP: 'readonly',
-	},
-}
