+++ lib/globals.d.ts
@@ -1,0 +1,6 @@
+declare global {
+	// Injected by vite
+	const __PACKAGE_VERSION__: string
+}
+
+export {}
