+++ package.json
@@ -1,48 +1,54 @@
 {
   "name": "@nextcloud/eslint-plugin",
-  "version": "2.2.1",
-  "license": "ISC",
+  "version": "3.0.0-beta.1",
   "description": "Nextcloud lint plugin for ESLint",
   "keywords": [
     "eslint",
-    "eslintplugin",
     "eslint-plugin",
     "nextcloud"
   ],
-  "author": "Christoph Wurst",
   "homepage": "https://github.com/nextcloud-libraries/eslint-plugin",
   "repository": "git+https://github.com/nextcloud-libraries/eslint-plugin.git",
-  "main": "lib/index.js",
-  "scripts": {
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage"
-  },
+  "license": "MIT",
+  "author": "Christoph Wurst",
   "files": [
     "docs",
-    "lib",
+    "dist",
     "CHANGELOG.md"
   ],
-  "peerDependencies": {
-    "eslint": ">=7.0.0"
+  "type": "module",
+  "main": "dist/index.cjs",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.cjs"
+    }
+  },
+  "scripts": {
+    "build": "vite --mode production build",
+    "test": "vitest run",
+    "test:coverage": "vitest run --coverage",
+    "test:watch": "vitest run --watch"
   },
   "dependencies": {
     "fast-xml-parser": "^4.5.1",
-    "requireindex": "^1.2.0",
     "semver": "^7.7.1"
   },
   "devDependencies": {
-    "eslint": "^8.46.0",
+    "@nextcloud/vite-config": "^2.2.2",
+    "eslint": "^9.14.0",
     "jest": "^29.7.0",
-    "memfs": "^4.17.0"
+    "memfs": "^4.17.0",
+    "typescript": "^5.6.3",
+    "vite": "^5.4.10",
+    "vitest": "^2.1.4"
+  },
+  "peerDependencies": {
+    "eslint": ">=9.0.0"
   },
   "engines": {
     "node": "^20.0.0",
     "npm": "^10.0.0"
-  },
-  "jest": {
-    "collectCoverageFrom": [
-      "lib/**/*.js"
-    ]
   }
 }
